refactor: fix pylint & tests.
