+++ service/routes/room.py
@@ -64,7 +64,7 @@ def create_group(auth_uid=None):
 
 @room.route('/room/<string:group_id>', methods=['GET'])
 @check_token
-def get_group_profile(auth_uid=None, group_id=None):
+def get_group_profile(group_id=None):
     """Get the profile of a matching group."""
     group_id = group_id or request.args.get('gid') or ''
     group_doc = ref_groups.document(group_id)
@@ -76,13 +76,11 @@ def get_group_profile(auth_uid=None, group_id=None):
     members = list_uid or []
     return_members = []
 
-    if Group.validate_user_role(auth_uid,
-                                group_id=group_id, group_snap=snap) > 0:
-        for user_id in members:
-            user_dict = ref_users.document(user_id).get().to_dict()
-            user_dict['userId'] = user_id
-            return_members.append(user_dict)
-        rv["members"] = return_members
+    for user_id in members:
+        user_dict = ref_users.document(user_id).get().to_dict()
+        user_dict['userId'] = user_id
+        return_members.append(user_dict)
+    rv["members"] = return_members
 
     return {
         "status": "success",
