+++ service/utils/connections.py
@@ -4,9 +4,9 @@ Initialize db collection reference.
 
 from google.cloud.firestore import CollectionReference
 
-from utils import db
+from utils import config_g
 
-ref_users: CollectionReference = db.collection(u'Users')
-ref_groups: CollectionReference = db.collection(u'Groups')
-ref_votes: CollectionReference = db.collection(u'Votings')
-ref_items: CollectionReference = db.collection("Items")
+ref_users: CollectionReference = config_g.db.collection(u'Users')
+ref_groups: CollectionReference = config_g.db.collection(u'Groups')
+ref_votes: CollectionReference = config_g.db.collection(u'Votings')
+ref_items: CollectionReference = config_g.db.collection("Items")
