+++ service/utils/decorators.py
@@ -6,7 +6,6 @@ from firebase_admin import auth
 from firebase_admin.auth import ExpiredIdTokenError, RevokedIdTokenError, InvalidIdTokenError
 from flask import request
 
-# pylint: disable=cyclic-import
 from utils import config_g
 from utils.exceptions import LoginRequired
 
@@ -26,7 +25,6 @@ def check_token(f):
             raise LoginRequired('No auth token provided')
         try:
             _, token_body = auth_token.split()
-            # if 'is_testing' in g or 'is_testing' in session:
             if config_g.is_testing:
                 user = {'user_id': token_body}
             else:
