--- service/utils/read_write.py
@@ -1,13 +1,0 @@
-"""
-read and write to file doc.
-"""
-import json
-
-
-def json_read(filename):
-    """
-    :param filename: file path
-    :return:
-    """
-    with open(filename, 'r') as file:
-        return json.load(file)
